requirements.txt
openai
faiss-cpu
fastapi
uvicorn
numpy
sentence-transformers